requirements_cloud.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core dependencies for cloud deployment
openai==1.75.0
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
pytest

# Additional dependencies
pydantic>=2.0.0 